+++ src/index.ts
@@ -1,3 +1,2 @@
+export {setup} from './setup';
 export {transform} from './transform';
-export {Reporter} from './reporter';
-export {create_runtime as setup} from './runtime';