+++ src/runtime.ts
@@ -1,6 +1,5 @@
-import {get_server_port_from_raw_config} from './config';
-import {Expressions, RawConfig, Snapshots} from './definitions';
-import {Server} from './server';
+import {Expressions, Snapshots} from './definitions';
+import {indent} from './utils/indent';
 
 export class Runtime {
 
@@ -20,25 +19,4 @@ export class Runtime {
     return `\nInferred\n\n${indent(this.expressions[line], 2)}\n\nto be\n\n${indent(this.snapshots[line], 2)}\n`;
   }
 
-}
-
-export const create_runtime = (
-    the_module: {filename: string},
-    config: RawConfig,
-    expressions: Expressions,
-    callback: (runtime: Runtime) => void) => {
-  const port = get_server_port_from_raw_config(config);
-  const {filename} = the_module;
-  const trigger_lines = Object.keys(expressions).map(line_str => +line_str);
-  Server.request_snapshots(port, filename, trigger_lines, snapshots => {
-    callback(new Runtime(expressions, snapshots));
-  });
-};
-
-function indent(str: string, spaces: number) {
-  let indentation = '';
-  for (let i = 0; i < spaces; i++) {
-    indentation += ' ';
-  }
-  return str.replace(/^/mg, indentation);
 }