+++ src/utils/remove-spaces.ts
@@ -1,0 +1,2 @@
+export const remove_spaces = (text: string) =>
+  text.replace(/^\s+|\s+$|\n/mg, '');