+++ CHANGELOG.md
@@ -10,6 +10,11 @@ All changes to this project will be documented in this file.
 > - 🏠 [Internal]
 > - 💅 [Polish]
 
+## Unreleased
+
+#### 💥 Breaking Change
+- Remove server since tests should be separated
+
 ## v20.0.6 (2017-06-10)
 
 #### 🐛 Bug Fix